requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.12
botocore==1.34.12
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kavenegar==1.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
s3transfer==0.10.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
